Consolidated EditableRadio into EditableInput.
