--- src/components/EditableRadio.tsx
@@ -1,68 +1,0 @@
-import * as React from "react";
-
-export interface EditableRadioProps extends React.HTMLProps<EditableRadio> {
-  label?: string;
-  name: string;
-  disabled: boolean;
-  onChange?: () => any;
-  checked: boolean;
-}
-
-export interface EditableRadioState {
-  checked: boolean;
-}
-
-/** Renders a radio button with a value that can be changed by the user.
-    Similar to `EditableInput`, but specifically for a radio button. */
-export default class EditableRadio extends React.Component<EditableRadioProps, EditableRadioState> {
-  constructor(props) {
-    super(props);
-    this.state = {
-      checked: props.checked || false
-    };
-    this.handleChange = this.handleChange.bind(this);
-  }
-
-  render() {
-    return (
-      <div className="radio">
-        <input
-          type="radio"
-          disabled={this.props.disabled}
-          name={this.props.name}
-          value={this.props.value}
-          ref="element"
-          onChange={this.handleChange}
-          style={this.props.style}
-          placeholder={this.props.placeholder}
-          checked={this.state.checked}
-          />
-        {" " + this.props.label}
-      </div>
-    );
-  }
-
-  componentWillReceiveProps(props) {
-    if (props.checked !== this.props.checked) {
-      this.setState({
-        checked: props.checked
-      });
-    }
-  }
-
-  handleChange() {
-    if (!this.props.onChange || this.props.onChange() !== false) {
-      this.setState({
-        checked: this.getChecked()
-      });
-    }
-  }
-
-  getChecked() {
-    return (this.refs as any).element.checked;
-  }
-
-  clear() {
-    this.setState({ checked: false });
-  }
-}
