+++ src/components/__tests__/EditableInput-test.tsx
@@ -32,6 +32,10 @@ describe("EditableInput", () => {
     wrapper.setProps({ "type": "checkbox" });
     label = wrapper.find("label");
     expect(label.text()).to.contain("label");
+
+    wrapper.setProps({ "type": "radio" });
+    label = wrapper.find("label");
+    expect(label.text()).to.contain(" label");
   });
 
   it("shows description from props", () => {
@@ -43,6 +47,12 @@ describe("EditableInput", () => {
     expect(wrapper.state().value).to.equal("initial value");
     let input = wrapper.find("input");
     expect(input.prop("value")).to.equal("initial value");
+    expect(input.prop("checked")).to.equal(true);
+  });
+
+  it("shows initial checked from props", () => {
+    expect(wrapper.state("checked")).to.equal(true);
+    let input = wrapper.find("input");
     expect(input.prop("checked")).to.equal(true);
   });
 
@@ -82,6 +92,23 @@ describe("EditableInput", () => {
       <EditableInput
         elementType="input"
         type="checkbox"
+        label="label"
+        name="name"
+        disabled={false}
+        checked={true}
+        />
+    );
+    input = wrapper.find("input");
+    inputElement = input.get(0) as any;
+    inputElement.checked = false;
+    input.simulate("change");
+    expect(wrapper.state().checked).to.equal(false);
+
+    // And a radio.
+    wrapper = mount(
+      <EditableInput
+        elementType="input"
+        type="radio"
         label="label"
         name="name"
         disabled={false}
