+++ src/stylesheets/custom_lists.scss
@@ -32,12 +32,11 @@
       }
     }
 
-    .radio {
+    .form-group {
       display: inline-block;
       margin-top: 15px;
       margin-left: 15px;
       margin-right: 25px;
-      font-size: 0.7em;
     }
 
     ul {
